% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{calc_prot_dissipation}
\alias{calc_prot_dissipation}
\title{Calculate real degradation rate, calculate dissipation rate}
\usage{
calc_prot_dissipation(prot_sum_data)
}
\arguments{
\item{prot_sum_data}{where does growth rate value come from?!}
}
\description{
Calculate the percent of the protein that was degraded during one generation(Step 12f-g in workflow, previously performed in Igor)
}
