% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{calculate_fraculab}
\alias{calculate_fraculab}
\title{Calculate labeled fraction 1}
\usage{
calculate_fraculab(data, meta_data)
}
\arguments{
\item{meta_data}{file?? should we write a separate function to read in metadata file?
Maybe ask Matt what format is most useful (what else is in metadata that they use?)}

\item{filtered_data}{the filtered svm data}
}
\description{
Calculate fraclab, add to metadata(replace excel step 11)
}
