% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{calculate_fraculab_clean}
\alias{calculate_fraculab_clean}
\title{Calculate labeled fraction and clean: added step to filter based on number timepoints where the peptide is identified}
\usage{
calculate_fraculab_clean(filtered_data, meta_data, min_timepoint_present = 3)
}
\arguments{
\item{filtered_data}{the filtered data}

\item{meta_data}{file?? should we write a separate function to read in metadata?}

\item{min}{number of timepoints present
another option is to make a separate function for filtering based on number of time points present}
}
\description{
Calculate fraclab, calculate number of occurences, add to metadata (Step 11 in workflow)
}
