% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_degradation.R
\name{calculate_label_rate}
\alias{calculate_label_rate}
\title{Calculate protein or peptide labeling rate}
\usage{
calculate_label_rate(data, combine_peptides = TRUE, quiet = FALSE)
}
\arguments{
\item{data}{the svm_data with fraclab/fraculab calculated \link{calculate_fraculab}}

\item{max_label}{amount of maximum label on scale 0-1 (1 = 100% label)}
}
\description{
calculate protein or peptide labeling rate from svm_data (fit each peptide to exponential equation y = Ae^dx)
}
