% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{filter_min_timepoints}
\alias{filter_min_timepoints}
\title{Filter data based on number of timepoints where the peptide is identified}
\usage{
filter_min_timepoints(data, min_timepoint_present = 3, quiet = FALSE)
}
\arguments{
\item{data}{the data to be filtered}

\item{min}{number of timepoints present}
}
\description{
(Step 11 in workflow)
}
