% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{filter_peptides_by_spectral_fit}
\alias{filter_peptides_by_spectral_fit}
\title{Process the SVM data}
\usage{
filter_peptides_by_spectral_fit(svm_data, pred_cutoff = 0.75, quiet = FALSE)
}
\arguments{
\item{svm_data}{the SVM data}

\item{pred_cutoff}{the probability cutoff from the svm prediction}
}
\description{
process and filter data (Step 10 in current workflow)
}
