% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{make_prot_master}
\alias{make_prot_master}
\title{make master list: combine peptide data to protein level}
\usage{
make_prot_master(deg_rate_data, min_chisq = 3, min_num_peptides = 2)
}
\description{
Average kdeg values of peptides per protein (Step 12d-e in workflow, previously performed in Igor)
}
