% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{plot_deg_curve}
\alias{plot_deg_curve}
\title{Plot degradation curves}
\usage{
plot_deg_curve(fraculab_data_clean, prots_to_plot = 10)
}
\arguments{
\item{fraculab_data}{or fraculab_data_clean}

\item{number}{of proteins (or peptides?) to plot?}
}
\description{
create graphs showing proportion of unlabeled fraction over time to visualize degradation
}
