% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{process_svm_data}
\alias{process_svm_data}
\title{Process the SVM data}
\usage{
process_svm_data(svm_data, pred_cutof = 0.75)
}
\arguments{
\item{svm_data}{the SVM data}

\item{pred_cutoff}{the probability cutoff from the svm prediction}
}
\description{
bla
}
