% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{read_svm_data_file}
\alias{read_svm_data_file}
\title{Read SVM data file}
\usage{
read_svm_data_file(filepath)
}
\arguments{
\item{filepath}{the path to the svm data file}
}
\value{
data_frame with the read data
}
\description{
This functions reads the data file
}
