% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tor_add_metadata}
\alias{tor_add_metadata}
\title{Add metadata}
\usage{
tor_add_metadata(data, metadata, join_by, quiet = FALSE)
}
\arguments{
\item{data}{the mass spec data}

\item{metadata}{a data frame of metadata}

\item{join_by}{which column to join by}
}
\description{
Add metadata
}
