% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{tor_add_protein_counts_info}
\alias{tor_add_protein_counts_info}
\title{Add protein counts info}
\usage{
tor_add_protein_counts_info(data, protein_counts, quiet = FALSE)
}
\arguments{
\item{data}{the data with the calculated labeling rates}

\item{protein_counts}{the protein counts data frame}
}
\description{
Add protein counts to the labeling rate data. This function adds the protein counts and only keeps proteins that have counts > 0 (with a warning). It also calculates relative abundances by counts and mass and thus requires the \link{tor_add_uniprot_info} function to have provided molecular masses of the proteins beforehand.
}
