% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_degradation.R
\name{tor_calculate_degradation_dissipation}
\alias{tor_calculate_degradation_dissipation}
\title{Calculate degradation rate}
\usage{
tor_calculate_degradation_dissipation(data, growth_rate, growth_rate_se = 0,
  quiet = FALSE)
}
\arguments{
\item{data}{data with label rate calculated}

\item{growth_rate}{growth rate from experiment}

\item{growth_rate_se}{standard error of growth rate from experiment}
}
\description{
Calculate degradataion and dissipation, including propagated error. Protein degradation is calculated as the labeling rate minus the growth rate (deg_rate = label_rate - growth_rate). Protein dissipation is calculated as the degradation rate divided by the labeling rate, and is displayed as a percent (dissipation = (deg_rate / label_rate)*100).
}
