% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_degradation.R
\name{tor_calculate_growth_params}
\alias{tor_calculate_growth_params}
\title{Calculate growth parameters}
\usage{
tor_calculate_growth_params(flow_rate, volume, flow_rate_se = NA_real_,
  volume_se = NA_real_)
}
\arguments{
\item{flow_rate}{the flow rate (mass/time)}

\item{volume}{the reactor volume (mass)}

\item{flow_rate_se}{the standard error of the flow rate (should be same units as flow_rate)}

\item{volume_se}{the standard error of the volume (same units as volume)}
}
\value{
a data frame with growth rate and its error as well as doubling time and its error
}
\description{
Calculate the generation time and growth rate in units of time and reciprocal time, respectively (same units as the denominator of the flow rate) as well as . The flow rate and volume must have the same mass units (e.g. g, kg, mL, or L). Calculates the standard errors as well if error estimates for flow rate and volume are provided.
}
