% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{tor_calculate_labeled_fraction}
\alias{tor_calculate_labeled_fraction}
\title{Calculate labeled fraction}
\usage{
tor_calculate_labeled_fraction(data)
}
\arguments{
\item{filtered_data}{the filtered svm data}
}
\description{
Calculate fraclab, add to metadata(replace excel step 11)
}
\details{
Cacluated the labeled and unlabaled fraction based on labeld and unlabeld amplitudes.
}
