% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{tor_fetch_kegg_details}
\alias{tor_fetch_kegg_details}
\title{Fetch KEGG details}
\usage{
tor_fetch_kegg_details(kegg_id, unnest_single_values = TRUE, ...,
  quiet = FALSE)
}
\arguments{
\item{kegg_id}{one or more KEGG identifiers. Can only do 10 requests at a time (see \link[KEGGREST]{keggGet}) so if there are more than 10, splits up into multiple queries.}

\item{unnest_single_values}{whether to unnest single values (values that have a none or only a single entry for all retrieve records)}
}
\description{
Helper function to convert output of \link[KEGGREST]{keggGet} to a nested data frame that is easy to work with in the tidyverse.
To unnest individual nested columns and preserve others use the \code{.drop=FALSE} parameter in \link[tidyr]{unnest}
}
