% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{tor_fetch_kegg_pathway_details}
\alias{tor_fetch_kegg_pathway_details}
\title{Fetch KEGG pathway details}
\usage{
tor_fetch_kegg_pathway_details(pathways)
}
\arguments{
\item{pathways}{data frame with column kegg_pathway_id}
}
\description{
Convenience wrapper to accept a pathways data frame and focus on the most relevant columns for pathways (uses \link{tor_fetch_kegg_details} internally).
}
