% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{tor_fetch_kegg_pathway_maps}
\alias{tor_fetch_kegg_pathway_maps}
\title{Fetch KEGG pathway maps}
\usage{
tor_fetch_kegg_pathway_maps(pathways, download_folder = "kegg",
  overwrite = FALSE, kgml = TRUE, species_png = TRUE,
  general_png = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{pathways}{data frame with columns kegg_org_id and kegg_pathway_id}

\item{download_folder}{which folder to download to (relative to working directory or absolute), will be created if it does not exist}

\item{overwrite}{whether to overwrite existing files}

\item{quiet}{parameter passed to \link[utils]{download.file}}

\item{...}{additional parameters passed on to \link[utils]{download.file}}
}
\value{
returns the unaltered pathways data frame
}
\description{
Downloads basic KEGG pathway maps
}
