% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{tor_fetch_kegg_pathways_for_species}
\alias{tor_fetch_kegg_pathways_for_species}
\title{Fetch all KEGG pathways for a species}
\usage{
tor_fetch_kegg_pathways_for_species(org_id = find_by_species_name(),
  species_name = NULL)
}
\arguments{
\item{org_id}{KEGG organism code (typically 3-4 characters), if not provided, will be searched based on \code{species_name} using \link{tor_fetch_kegg_species}}

\item{species_name}{species name or regular expression to identify the organism of interest}
}
\description{
Fetch all KEGG pathways for a species
}
