% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{tor_fetch_uniprot_proteins}
\alias{tor_fetch_uniprot_proteins}
\title{Fetch uniprot information}
\usage{
tor_fetch_uniprot_proteins(taxon = 83333, read_cache = TRUE,
  cache_dir = "cache")
}
\description{
Fetch uniprot information
}
