% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{tor_fetch_uniprot_species}
\alias{tor_fetch_uniprot_species}
\title{Fetch uniprot species}
\usage{
tor_fetch_uniprot_species(search_term, quiet = FALSE)
}
\arguments{
\item{search_term}{taxa search term}
}
\description{
Searches the uniprot data base for bacterial species based on the seacrh term.
}
