% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tor_filter_label_rate_fits}
\alias{tor_filter_label_rate_fits}
\title{Filter label rate fits}
\usage{
tor_filter_label_rate_fits(dt, condition = TRUE, select = everything(),
  quiet = FALSE)
}
\arguments{
\item{dt}{data label rates calculated}

\item{condition}{filtering condition. Includes all records by default.}

\item{select}{which columns to select - use \code{c(...)} to select multiple, supports all \link[dplyr]{select} syntax including renaming columns. Includes all columns by default.}
}
\description{
Filter protein or peptide label rate fit records for quality, visualization or focus on problematic records. This function is typically called some time after \link{tor_calculate_label_rate} or \link{tor_calculate_degradation_dissipation} before processing the label rate fits further. Use the \code{condition} parameter to set filtering conditions. Use the \code{select} parameter to keep only the most informative columns.
}
