% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{tor_filter_peptides_by_spectral_fit_quality}
\alias{tor_filter_peptides_by_spectral_fit_quality}
\title{Filter peptides}
\usage{
tor_filter_peptides_by_spectral_fit_quality(svm_data, condition,
  quiet = FALSE)
}
\arguments{
\item{condition}{the filtering condition}

\item{data}{the data set}
}
\description{
Process and filter data (Step 10 in current workflow) by spectral quality evaluation.
}
