% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tor_plot_label_rate_error}
\alias{tor_plot_label_rate_error}
\title{plot label rate error}
\usage{
tor_plot_label_rate_error(data)
}
\arguments{
\item{data}{data with label rate calculated, including growth rate standard errors}
}
\description{
visualize the label rate and residual standard error
}
