% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tor_plot_labeling_curves}
\alias{tor_plot_labeling_curves}
\title{plot disippation curves}
\usage{
tor_plot_labeling_curves(data, grouping = protein, plot_number = 6)
}
\arguments{
\item{data}{data with label rate (and error), growth rate (and error)}

\item{grouping}{by protein or peptide for visualization ??}

\item{plot_number}{designate number of curves you want to see}
}
\description{
visualuze the dissipation curve of peptides or proteins
}
