% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{tor_read_protein_counts_data}
\alias{tor_read_protein_counts_data}
\title{Read protein counts data}
\usage{
tor_read_protein_counts_data(psms_file_path, quiet = FALSE)
}
\arguments{
\item{psms_file_path}{the path to the protein counts csv file - expected columns are \code{protein} (the protein ID) and \code{T0_counts} - the protein counts at the T0 condition.}
}
\description{
Reads the protein counts data from a `psms.csv` file. Note that the columns from the file get slightly renamed for compatibility with other functions.
}
