% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{tor_read_svm_data_file}
\alias{tor_read_svm_data_file}
\title{Read SVM data file}
\usage{
tor_read_svm_data_file(filepath, quiet = FALSE)
}
\arguments{
\item{filepath}{the path to the svm data file}
}
\value{
data_frame with the read data
}
\description{
This functions reads SVM preprocessed data files and selects the most important information
}
