% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{tor_recode_protein_ids}
\alias{tor_recode_protein_ids}
\title{Rename proteins based on protein mapping file}
\usage{
tor_recode_protein_ids(data, renaming_protein_map_file, prot_col = "prot_id",
  prot_new_col = "new_prot_id", quiet = FALSE)
}
\arguments{
\item{data}{the data set}

\item{renaming_protein_map_file}{the filepath to the xlsx mapping file}

\item{prot_col}{the name of the column that has the protein ID/name}
}
\description{
This function may be obsolete once uniprot integration is completed.
}
