% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tor_remove_list_columns}
\alias{tor_remove_list_columns}
\title{Remove complex data}
\usage{
tor_remove_list_columns(dt)
}
\description{
Convenience function to remove nested and list columns in a data table (e.g. in preparation for printing to console or RMarkdown).
}
