% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{tor_rename_proteins}
\alias{tor_rename_proteins}
\title{Rename proteins based on protein mapping file}
\usage{
tor_rename_proteins(data, renaming_protein_map_file, prot_col = "prot",
  prot_new_col = "protNew", quiet = FALSE)
}
\arguments{
\item{data}{the data set}

\item{renaming_protein_map_file}{the filepath to the xlsx mapping file}

\item{prot_col}{the name of the column that has the protein ID/name}
}
\description{
This function may be obsolete once uniprot integration is completed.
}
