% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{tor_show_package_structure}
\alias{tor_show_package_structure}
\title{Generate flow chart}
\usage{
tor_show_package_structure(height = NULL, width = height, ...)
}
\arguments{
\item{height}{height of the chart}

\item{width}{width of the chart}

\item{...}{passed on to \link[DiagrammeR]{mermaid}}
}
\description{
To render icons correctly in RMarkdown files, make sure to include \code{htmltools::tagList(rmarkdown::html_dependency_font_awesome())}.
}
