% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{turnoveR-package}
\alias{turnoveR}
\alias{turnoveR-package}
\title{turnoveR: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/turnoveR}
  \item Report bugs at \url{https://github.com/kopflab/turnoveR/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
