% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_degradation.R
\name{calc_pep_degrate}
\alias{calc_pep_degrate}
\title{Calculate degradation rate, chisquared value}
\usage{
calc_pep_degrate(fraculab_data_clean)
}
\arguments{
\item{fraculab_data}{or fraculab_data_clean}
}
\description{
calculate kdeg and chisquared (Step 12a-c in workflow, previously performed in Igor)
}
