% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{calculate_fraculab}
\alias{calculate_fraculab}
\title{Calculate labeled fraction 1}
\usage{
calculate_fraculab(data)
}
\arguments{
\item{filtered_data}{the filtered svm data}
}
\description{
Calculate fraclab, add to metadata(replace excel step 11)
}
