% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_degradation.R
\name{tor_calculate_degradation_dissipation}
\alias{tor_calculate_degradation_dissipation}
\title{Calculate degradation rate}
\usage{
tor_calculate_degradation_dissipation(data, growth_rate, growth_rate_se = 0)
}
\arguments{
\item{data}{data with label rate calculated}

\item{growth_rate}{growth rate from experiment}

\item{growth_rate_se}{standard error of growth rate from experiment}
}
\description{
Calculate degradataion and dissipation, including propagated error.
}
