% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_data.R
\name{tor_calculate_labeled_fraction}
\alias{tor_calculate_labeled_fraction}
\title{Calculate labeled fraction}
\usage{
tor_calculate_labeled_fraction(data, unlabeled_col = "amp_ulab",
  labeled_col = "amp_lab", quiet = FALSE)
}
\arguments{
\item{data}{the dataset}

\item{unlabeled_col}{the unlabeled singal column}

\item{labeled_col}{the labeled signal column}
}
\description{
Cacluated the labeled and unlabaled fraction based on labeld and unlabeld amplitudes.
}
