% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{tor_fetch_uniprot_proteins}
\alias{tor_fetch_uniprot_proteins}
\title{Fetch uniprot information}
\usage{
tor_fetch_uniprot_proteins(taxon, read_cache = TRUE, cache_dir = "cache",
  quiet = FALSE)
}
\arguments{
\item{taxon}{the ID of the species}
}
\description{
Fetch uniprot information
}
