% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tor_plot_label_rate_error}
\alias{tor_plot_label_rate_error}
\title{Plot label rate error}
\usage{
tor_plot_label_rate_error(data)
}
\arguments{
\item{data}{data with label rate calculated, including growth rate standard errors}
}
\description{
Visualize the label rate and residual standard error. Displays the residual standard error (i.e. the normalized error of the fit) for each calculated label rate to easily evaluate what quality cutoffs might be useful. Plots the growth rate as a reference and the individual standard errors of the parameters from the fit. Supports plotting of multiple datasets by coloring according to growth rate.
}
