% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tor_plot_label_rate_hist}
\alias{tor_plot_label_rate_hist}
\title{Plot label rate histogram}
\usage{
tor_plot_label_rate_hist(data)
}
\arguments{
\item{data}{data with label rate calculated, including growth rate}
}
\description{
creates histogram of labeing rate
}
