% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{turnoveR-package}
\alias{turnoveR}
\alias{turnoveR-package}
\title{turnoveR: Processing tools for protein degradation experiments.}
\description{
This packages provides a collection of data processing, protein information and data visualization funtions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/turnoveR}
  \item Report bugs at \url{https://github.com/kopflab/turnoveR/issues}
}

}
\author{
\strong{Maintainer}: Corinne Walsh \email{corinne.walsh@colorado.edu}

Authors:
\itemize{
  \item Sebastian Kopf \email{sebastian.kopf@colorado.edu}
}

}
\keyword{internal}
